move world definitions to separate file
